+++ frontend/src/layouts/cameras/components/DeviceSwitch.tsx
