+++ frontend/src/layouts/lights/types.tsx
@@ -1,0 +1,12 @@
+export interface LightDevice {
+    pin: number;
+    nickname: string;
+    controller_index: number;
+    controller_name: string;
+    intensity: number;
+}
+
+export interface PWMController {
+    name: string;
+    pins: number[];
+}
