+++ frontend/src/layouts/logs_page/types.tsx
@@ -1,0 +1,9 @@
+export interface Log {
+    timestamp: string;
+    level: string;
+    name: string;
+    filename: string;
+    lineno: number;
+    function: string;
+    message: string;
+}
