+++ frontend/src/layouts/preferences/index.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { Grid, Typography, TextField, Box, Button } from "@mui/material";
-import { getSettings, savePreferences } from "../../utils/api";
+import { getSettings, savePreferences } from "./api";
 import { styles } from "../../style";
 import { IP_REGEX } from "../../utils/utils";
-import SettingsCard from "../../components/SettingsCard";
-import TextFieldButton from "../../components/TextFieldButton";
+import SettingsCard from "./components/SettingsCard";
+import TextFieldButton from "./components/TextFieldButton";
 import { useSnackbar } from "notistack";
 
 const PreferencesLayout = () => {
