+++ frontend/src/layouts/preferences/types.tsx
@@ -1,0 +1,5 @@
+import { StreamEndpoint } from "../cameras/types";
+
+export interface SavedPreferences {
+    default_stream: StreamEndpoint;
+}
