+++ frontend/src/layouts/updater/types.tsx
@@ -1,0 +1,16 @@
+/* Updater */
+
+export interface Release {
+    body: string;
+    name: string;
+    prerelease: boolean;
+    published_at: string;
+    tag_name: string;
+    url: string;
+    current: boolean;
+    mostRecent: boolean;
+}
+
+export interface ReleaseList {
+    releases: Release[];
+}
